Improved NotificationService (still not working properly, the services shut downs when the application closes, and there is a error when we delete accounts)
Using share preferences to have a battery safe mode.

If battery safe mode the service is not been initialize for the notifications.

The NotificationService is giving a "debug" notification that is been turned on.

The map API is not been shown when the user doesnt have any WIFI

Now we can delete users in the setting section. Pictures, rooms, accounts
are deleted in the process.